+++ packages/icons/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## Unreleased
 
+### Automated Tests
+- Added test to make sure smart search data for icon libraries are update always.
+
+
 ## 1.0.0 (2024-12-08)